Fix CI build error
 - Now the docker image should build with an extant build directory and a new one
 - There may be some subtle issues about permissions and directory creation on
   Travis-CI that still exist


Former-commit-id: 48dc12a6d7caee8acaad3cf7408098cd99f980c2

+++ travis.sh
@@ -12,8 +12,10 @@ docker run -i --rm -u vagrant -v "$PWD":/home/vagrant/src/radiasoft/zgoubi "${1:
     source ~/.bashrc
     set -euo pipefail
     cd ~/src/radiasoft/zgoubi
-    mkdir build > /dev/null 2>&1 || echo \
-          "Warning: Using an old/dirty build directory! Please run 'rm -r build' and try again if script fails." >&2
+    if ! mkdir build > /dev/null 2>&1 ; then
+        echo \
+            "Warning: Using an old/dirty build directory! Please run 'rm -r build' and try again if script fails." >&2
+    fi
     cd build
     cmake -Wdev -DCMAKE_INSTALL_PREFIX=$(pyenv prefix) ..
     make -j $(nproc)
